+++ src/lua/api/waypoints.lua
@@ -2,7 +2,7 @@ local raw = trxc.waypoints
 local api = trx.api
 
 api.module("waypoints", {
-  order = 36,
+  order = 37,
   description = [[
     Module for how far along a level's own progression Lara has got.